src/requirements.txt
# --- Core Backend ---
flask==3.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.29

# --- MySQL ---
mysql-connector-python==8.3.0

# --- LangChain Stack ---
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
langchain-groq==0.0.1

# --- Optional (only needed if using frontend)
